Add files via upload
Proud to be nurtured by the National Security Agency as a foreign citizen and very proud to be an MITx certificate holder, Harvard CCE certificate holder. This may be finished by this afternoon, 5pm 1st Sept 2024, two days after International Day of Commemoration for Forced Disappearances. May leave the Bay or migrate overseas in the next few years